pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.7",
    "dotenv>=0.9.9",
    "httpx-socks>=0.11.0",
    "httpx[socks]>=0.27",
    "langchain>=1.2.0",
    "langchain-nebius>=0.1.3",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.4.11",
    "langgraph-prebuilt>=1.0.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-telegram-bot>=22.5",
]
